Allow for env variables in the helm release name
When using skaffold in CI, it would be handy to be able to deploy
to a shared cluster. To do that the current options are either to
change skaffold.yaml during the CI job (e.g. using sed) or to limit
CI jobs to one at the time for a specific repo/skaffold.yaml.

Templating the release name allows to bind the release name to
change metadata like a PR number.

Adding tests for util/env_template.go which I prepared for refactor
PR and missed to add into the PR.

Signed-off-by: Andrea Frittoli (andreaf) <andrea.frittoli@uk.ibm.com>