Implement packaging for helm deployment
This commit adds a possibility to package a helm chart before deployment
using `helm package` command. This can be useful to set a specific version
for a chart.

---

One of the workflows that skaffold assumes is to deploy a Helm chart that
resides within some directory in the project folder. `Chart.yaml` file in
that folder is basically static file, and `--version` flag in
`helm install ./path/to/chart --version` command actually does nothing,
so does "version:" field in helm section of **skaffold.yaml**.

To set a chart version dynamically, we have to use `helm package` command
setting `--version` (and optionally `--app-version`) flags accordingly.

This commit adds a new section `packaged:` to helm section **skaffold.yaml**
that can be used to configure dynamic chart versions.