Fail if python script not found and run from any dir
If the script was run from anywhere other than the directory above hack,
the python script could not be found, however since the error code was
ignored, this would pass.

Finding files and ignoring failed files was being done in one line. We
wanted to ignore errors from grep, however this has the side effect of
ignoring errors from the python script, e.g. if this script is run
from an unexpected directory and the python script is not found.

Lastly, filtering of dirs was being done in both `boilerplate.sh` and
`boilerplate.py` so now it will just be done in `boilerplate.sh`.