Add a pop of color to terminal output
This commit breaks up the Skaffold terminal output with a pop of color!

Skaffold draws a distinction between log output, which goes to stdout,
and output from Skaffold to the user, which goes to stdout and is
interleaved with output from the tools Skaffold users (e.g. docker for
building). This commit changes the output from Skaffold to the user to
be Blue!

The escape codes to make the text blue (or other colors, in the case of
the kube pod-specific logging) will only be output if stdout is a
terminal. If output is being redirected to a file for example, the
escape codes will not be included.

This change makes use of the exciting `colors.go` logic but splits it
into the picker (specifically for mapping images to colors consistently)
from the logic to add escape codes.

Possible alternatives/improvements:
* Pass the formatter object around directly instead of `io.Writer`. One
  big downside to this is it would make it harder for callers to control
  the color of the output (right now it can be arbitrarily changed
  anywhere, e.g. `fix.go` outputs both blue text and plain text)
* Split the formatting from the output.

Fixes #835