Implement terminal color using wrapper around io.Writer
In #850 I proposed an implementation of terminal color which involves
instantiating a formatter object whenever you want to log to the
terminal in color. This PR instead creates a wrapper around io.Writer
instead, so that theoretically everything output to the terminal will
magically be the right color without callers needing to be aware.

That was how it was in the first iteration HOWEVER when I ran `skaffold
dev` _everything_ was blue, including output from docker, etc. It became
clear that we needed to often use the output writer which didn't apply
formatting, so in the end we need to change pretty much every function
that took io.Writer to take color.Writer instead, so we can control
whether the functions they call output in color or not.

The end result is that when you look at some code, if it still takes
io.Writer as a param, it's very hard to know if that will be in color or
not and I feel this implementation is more confusing and less
maintainable than #850.

(This was a possible fix for #835.)