Add DEVELOPMENT.md
This DEVELOPMENT.md will hopefully help folks ramp up and get started
contributing to skaffold by describing the development process.

After discussions with @r2d4 and @balopat we want to try out keeping the
contributor/developer docs in markdown, while the user facing docs will
continue to be in asciidoc. One of the reasons for this is that github
will automatically link to the CONTRIBUTING.md and code of conduct, and
so if we wrote this in ascii doc, we would _have_ to make the docs (and
any references, links, etc.) that they contain work for both markdown
view in github and ascii doctor view in the generated docs.