Add automatic port-forwarding to skaffold dev
Skaffold dev will now automatically port-forward any pods that expose
container ports.

Ports are assigned on a first-come-first-serve basis,
and that container will hold the lock on that port for the remainder of the dev
cycle.

Updates to a container will trigger an update which replaces the
forwarded port with the newest version.

Scaling up a deployment will continue to port-forward the latest pod.
However, scaling down a deployment has the potential to kill the
port-forwarded pod without replacement - since we do not keep track of
candidates that are "waiting" on that port.