Add a recreatePods option to helm releases
It is common in development to just rebuild a Docker image with the
same tag, which will not trigger a rolling update of the deployment
because the pod spec does not change. Passing --recreate-pods will
forcibly terminate the running pods and allow new ones to be
spawned with the new image.