Better API version parsing.
This implements part of the blang/semver surface for k8s-style group/versions.

I couldn't find an easy-to-use version of this in the k8s codebase, but I'm sure it's somewhere.
This gives us a way to tell if the unequal version is higher or lower, so we can give a better error message.