Add skeleton for webhook code
This is the skeleton for the code which will interact with the github webhook, allowing container-tools-bot to comment on PRs with changes to the docs with a URL to demonstrate those changes. I added a Dockerfile to build the docs controller image and the kubernetes yaml I ran in cluster to give the controller permissions to interact with the cluster.
I also added a Makefile rule to build the image and code for starting the server.