Fixed panic if skaffold.yaml is empty (#1216)
The absence of apiVersion from skaffold.yaml hadn't been tested, a
nil err was being sent to errors.Wrapf which then returned nil as error
leading to wrong success of configuration parse operation.

We are now creating a new error instead of relying on wraping a
potentially nil err with errors.Wrapf.