Create deployment to serve docs website via hugo
Adds support for creating a deployment which will map to a service with
an external IP.

The deployment consists of an init container and a regular container.
The init container is responsible for copying the PR branch (with docs
changes) into an emptyDir volume. The volume is then mounted into a
container, which runs `hugo server -D --bind 0.0.0.0 --baseURL
external-service-ip`. Once this deployment stabilizes, the static
website reflecting changes to the docs should be availble at the
external IP address associated with the service.

Next steps include:

1. Commenting on the Github PR once the deployment is ready
2. Uploading logs to GCS and commenting on the Github PR with a link to
those logs if the deployment should fail