Get builder, tagger and deployer info from config
This change should help with two things:
  1. metrics collection
  2. the "skaffold status" command

For metrics, this will alow us to get data for each version of the
skaffold config (for example, which builders are most prevalent in v1alpha2
etc).

For "skaffold status" we can get information about the config from the
config itself and display it for the user.