Create directory before kubectl cp
This PR should fix a bug, where syncing files into new directories wasn't working because those directories didn't exist. This change will create the directory in the container via kubectl exec. This change also requires that the container has 'mkdir' in it, but I couldn't think of a better way to get around this.

Also adds an integration test for file sync.