Fix labeling cluster-wide resources
This commit fixes the issue when skaffold labeler component tries to
set labels on cluster-wide resource.

Because PATCH was applied through a specific namespace of deploy
artifact, it always failed on cluster-wide resources like ClusterRole,
ClusterRoleBinding and others with an error "WARN[xxxx] error adding
label to runtime object: patching resource namespace-name/resource-name:
the server could not find the requested resource". Changes in this commit
makes labeler check if a resource is namespaced or cluster-wide.

Signed-off-by: Igor Zibarev <zibarev.i@gmail.com>