Add artifact cache to track artifacts for faster restart
I added an artifact cache (stored in ~/.skaffold/cache) which maps a
hash of the artifact's dependencies : image name and image tag. I used
kaniko's snapshotting library, which does exactly this for kaniko
caching.

Before building an artifact, we check the cache to see if an image with
the same hash and image name has already been built. If it has, we skip
rebuilding this image. If either the hash or the image name is
different, skaffold will rebuild the image.

The cache can be disabled with the --cache-artifacts=false flag.

Future work could include optimization around retagging cached images
(if we find the hash for an image, but the image name is different,
perhaps we can try to retag the image instead of rebuilding).