Fix unsupported transitive target case
When a Dockerfile is built and a target is provided,
Skaffold needs to find the dependencies for that target.

The code was missing an important case of transitive
dependency when a target `A` depends on a target `B`
because it has a `COPY --from=B` instruction.
(Could also be a `COPY --from=1` assuming target `B`
is the second target)

Fixes #1605

Signed-off-by: David Gageot <david@gageot.net>