Remove regression from port forwarding
I introduced this regression when adding "pod name" to they port forward
entry key -- whenever a pod was replaced, it's name would change and it
would be mapped to a new port.

This should fix that issue, but I initally introduced "pod name" so that
replicated pods would map to different ports. I'll work on finding a
solution for that in another PR.