Store workspace-hash:image-details, where details contains both image ID and digest
When using the cache with minikube, there were no image digests
available since the images had not been pushed to and pulled from a
registry. To get around this, I store image ID as well and search for a
local match to both the digest and the ID.

I also check if running against a local cluster, and only push cached
images if not.