Store core plugins within skaffold binary
If a user tries to use a core plugin, we call skaffold again with the
environment variable SKAFFOLD_PLUGIN_KEY=42 set, and with
SKAFFOLD_PLUGIN_NAME=[plugin name] set. If a user tries to use a plugin
that isn't a core plugin, skaffold will call out to the binary as it did
before.