Fix regression in sync
In #1709 I introduced a regression to sync by removing the "-C /" from
the tar extraction, which forced all files to be extracted at root. We
actually still want this, since now that we append the workdir for relative paths  all paths in the
tarball are absolute paths and so should be extracted at root.

This should fix #1721.