Resolve dependency files together with a destination list
Before, only a list of dependencies relative to the workingdir on the host
system was returned. Now, return a map of paths on the host system with
a list of target destinations in the container. This makes only really
sense for Dockerfile based containers, but allows to infer the file
locations that need to be updated when a file on the host changes.
This commit only changes the data structure and adapts tests where
necessary.

Signed-off-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>