Remove KanikoBuild type
This commit removes the KanikoBuild type, which is confusing as it
reprseents both a builder and an environment. It also assumes that
certain elements (buildcontext) apply to all artifacts, when in reality
different artifacts should have different buildcontext.

KanikoBuild is replaced with a KanikoArtifact type, which can be
specified per artifact. I also added a ClusterDetails field to contain
info about the cluster itself (namespace, timeouts, secrets etc)