Refactor KanikoBuild into KanikoArtifact and Cluster (#1797)
* Added KanikoArtifact type, and changed GetDependencies for Dockerfile
so that other artifacts can use the function.
* Remove KanikoBuild type
This commit removes the KanikoBuild type, which is confusing as it
reprseents both a builder and an environment. It also assumes that
certain elements (buildcontext) apply to all artifacts, when in reality
different artifacts should have different buildcontext.

KanikoBuild is replaced with a KanikoArtifact type, which can be
specified per artifact. I also added a ClusterDetails field to contain
info about the cluster itself (namespace, timeouts, secrets etc)
* Add integration test which builds two artifacts with kaniko
* Boilerplate, linting, and add integration test to run_test.go
* Review comments
* update skaffold.yaml for kaniko sub folder integration test to match new schema