Improved pipeline config for artifact.sync
Currently, the pipeline config is a map of local glob pattern to destination directory.
This scheme has been extended with some magic sequences, making it difficult to understand.
This commit converts the sync map into a list of sync rules. All sync rules are consulted to determine the destination paths.

This prepares for further changes to the sync logic. Also see #1844

Signed-off-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>