Refactor pod selection logic for logging
So far, pods were selected based on known image names in their container spec. This is restricting, because it does not allow to add additional pods for log-tailing, nor is it feasible to exclude uninteresting pods. Besides, the book-keeping code for the deployed images was scattered in the code.
Now, pods are selected for log-tailing by looking for a `tail=true` label on the pods. This will also allow to implement more interesting label selectors in the future.

Signed-off-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>