super hacky jkcfg proof-of-concept
jkcfg is a data templating tool that uses plain old javascript.   Like kustomize
it's hermetic.

this PR is literally a single line change which runs `jk run <dir>/install.js`
instead of `kustomize build <dir>`.   Obviously not ready for merging, but it's
enough to actually be useful.

To use it, use the kustomize clause in your skaffold.yaml:

    deploy:
      kustomize:
        path: jk_dir

put an kustomization.yaml file in that directory.  It's fine if it's completely
empty, but skaffold will watch any files you list in the resources section.  The
files don't have to be valid for kustomize.

name your root jkcfg file 'install.js'.

install.js must write out kubernetes config files to stdout:

    std.write(k8s_manifests, '', {format: std.Formatl.YAMLStream});