Refactor kaniko builder to cluster builder
Custom artifacts will need to work with both the local builder and the
cluster builder. However, currently the "cluster" builder is really just
the kaniko builder.

This refactor renames `pkg/build/kaniko` to `pkg/build/cluster`, and
reorganizes it so that it looks like `pkg/build/local`. This way, we
should easily be able to add a new CustomArtifact to the cluster
builder just as easily as we can to the local builder.