Check for env variables for root cmd persistent flags
We were checking for environment variables for subcommands of
`skaffold`, but not the global flags that directly apply to the `skaffold`
root cmd.

This should allow us to also specify `SKAFFOLD_FORCE_COLORS` as an env
variable, fixing #2139.

I wasn't sure how to unit test this without changing the environment
variable locally, and I didn't want to risk messing up anyone's
configuration. However, I tested locally and the env variable seems to
be working now.