Fix(#2290): implement checksum computation for sha256 tagger
Instead of returning a "latest" tag by default the tagger will now
return a checksum based on the sha256 of the working dir.

The function walk across all files and directories of the working dir,
sum the sha256 of each and compute a checksum of the resulting sum.

Errors on some files only are ignored. If the directory cannot be walk
at all, then the default tag is latest.