Feat(sync): skip sync on non-running pods (#2403)
Fix issue #2360

The goal of this PR is to avoid trying to sync files in a pod which is
not in a running state. It should avoid errors when working on a single
artifact which is deployed in different pods. For example : one running
a webserver and another running a work or a job. This is typically
something that could happend when a microservice is able to run both web
server and a job which make some sort of batch treatments on the database.