Fix bug in port forwarding
Port forwarding wasn't being restarted after each loop of the dev cycle,
so I've moved it to happen after runner.Deploy is called. This PR also
fixes a bug where used local ports were being stored as type int but not
deleted because the code was trying to look up the keys as type int32.
I ammended a unit test to make sure local ports are cleaned up correctly
as well.

Note: Ports are forwarded after each loop of the dev cycle, but they
aren't being forwarded to the same port.

So, on loop 1, svc/mysvc will be forwarded to port 9000.
Then, on loop 2, svc/mysvc will be forwarded to port 9001.

Currently, port 9000 is coming back as "in use" even though it's been
removed from the map of local ports and the go function calling it has
been cancelled. I'll look into this in another PR.