Fix issue with remote Kustomizations in dev mode. (#2581)
Remote Kustomization support was added previously, but it only supported
remote resources referenced inside of a Kustomization config file rather
than remote Kustomization config files themselves. This commit causes
the file watcher code to assume the absence of a Kustomization config
means that it is a remote reference and skips it since it can't watch
remote files for changes. It pushes error handling onto Kustomize
itself for invalid remote locations.