Jib Auto Sync - core + maven implementation
So most of this is baked into the jib package, but I don't think
anything in it necessarily HAS to be jib specific. If we want to
extend the model out for other systems we can. But it mostly
feels like a java problem anyway?

Gradle is not covered here.