Add `Suggestion` to event API.
In this PR,
1. Add `Suggestion` to the `ErrDef`. `Suggestion` represent an action for a user to perform to recover from an error.
  Right now `Suggestion` includes:
  - `SuggestionCode` uniquely identifies a suggestion.
  - `Action` describes the action or step to recover.
  In future it can contain some metadata like if this message is kubernetes specific etc.

2. Mark `Err` and `StatusCode` depreacted in `BuildEvent`, `DeployEvent`, `StatusCheckEvent`,
   `ResourceStatusCheckEvent`, `FileSyncEvent` where these fields were used.
  Add a new field `ActionableErr` which contains
   - ErrCode: distinctly identifies an error
   - Message: string describing error message
   - newly added `Suggestion`
This is to ensure, `ActionableErr` is nil if no error happens

See this on why, proto3 defauls fields not set to their default value
```
In proto3, all fields are "optional" (in that it is not an error if the sender fails to set them). But,
fields are no longer "nullable", in that there's no way to tell the difference between a field being
explicitly set to its default value vs. not having been set at all.
```