Use alternative service port-forwarding scheme (#4590)
When Skaffold redeploys new workload resources, it tears down and
re-establishes port-forwards for services.  But `kubectl port-forward
service/xxx` actually selects and forwards to a matching pod, in a
way that favours older pods -- pods that are often about to be
replaced during the rollout.  We select a pod matching the service
but favour newer pods which are more likely to be from the rollout.