Make the dev loop interruptible.
Wether it’s during the build, the deploy or the
status check, the dev loop can now be interrupted
on a file change.

Fix #3908
Fix #916

If it’s the skaffold.yaml that’s modified, instead
Of restarting the dev loop right away, we first
validate that the config is valid.

Fix #4225

Signed-off-by: David Gageot <david@gageot.net>