Update Paketo buildpack references
The `paketo-buildpacks/go-compiler` buildpack was renamed to `paketo-buildpacks/go-dist`. This `go-dist` buildpack will be sufficient to identify a Go application as it would be involved in any build of a Go application.

Additionally, it makes more sense to identify the project as Paketo. Its affiliation is with the Cloud Foundry foundation, not the project.