[v2] Remove the latest/v2 schema dependencies.
latest/v2 used to depend on latest/v1 for build, tag, test. This is aimed at making sure the v1 changes on these steps
can be backportting to v2. This requires engineers to update the v2 reference (e.g. bumps the latestV1 reference to a newer release)
and sometimes change the v2 pipeline as well. Since v2 is not released yet, this work are not officialized and v2 always fells behind
the v1 changes.

This change makes latest/v2 no longe relies on v1 schema.