Add --wait=false flag to skaffold cleanup kubectl delete to speedup the phase
What is the problem being solved?
Fixes #6123, Skaffold process takes a long time to terminate after ctrl-c. This PR:
- adds `--wait=false` flag to kubectl delete that is used during skaffold's Cleanup phase

Why is this the best approach?
This is the best approach because it makes the cleanup phase take considerably less time which is good from a UX perpective and more in line with what a user expects when sending a SIGTERM (Ctl-c).  Before this change the Cleanup phase could take from ~800ms -> ~10s.  With this change, the Cleanup phase is consistently ~600ms for the same app and skaffold.yaml.

What other approaches did you consider?
The alternative approach would be to add a line to stdout that explain to the user what skaffold is deleting items which can sometimes take a bit (vs making this change).  Currently w/o no explanation skaffold hanging can be a bit confusing.
.
What side effects will this approach have?
If users rely on skaffold waiting for the kubectl delete to finish, then this approach might alter/break those use cases.

What future work remains to be done?
N/A.