Add core ko builder implementation (#6054)
* Add core ko builder implementation

This adds the `Build()` method for building artifacts using ko. It
supports both publishing the resulting image to a registry, and
sideloading it to the local Docker daemon.

The Skaffold docker client is used in the ko builder. This ensures that
any Minikube config set by Skaffold is used. The ko builder uses the
docker client when sideloading images to the docker daemon.

The `temporary.go` file contains the structs intended to be added to the
schema.

The additions to the design proposal explain image naming for the ko
builder, specifically how container images are named and how Go import
paths are resolved when using the proposed ko builder.

This commit includes ko builder unit tests for
non-current-working-directory workspace dirs. These verify
that the ko builder works even if the context specified in
`skaffold.yaml` differs from the current working directory.

This implementation is still missing the following features:

- integration test
- dependencies (for file watching)
- insecure registries
- debug mode
- support for `go` flags and environment variables (based on
  https://github.com/google/ko/pull/340)
- actually plumbing the builder into the Skaffold CLI and API :-)

Tracking: #6041

* Improve ko Builder unit tests

Split out the confusing end-to-end unit test with tests for individual
functions.

* Reorder test inputs and expected outputs

Also fix typo in design proposal.

* Add TODO to update import path

Reminders to update the import path once the contents of
`pkg/skaffold/build/ko/schema` have been added to the real schema in
`pkg/skaffold/schema/latest/v1`.

* Fix lint errors for TODO comments in import block

Adding the TODO in the import block resulted in gci linter errors:
https://app.travis-ci.com/github/GoogleContainerTools/skaffold/jobs/530642848#L313