fix: TestDebug/buildpacks by pinning go version
This temporarily pins the Go version used by the buildpack-built image
used for testing in the "skaffold debug" integration test.

It does not fix the problem for the users (i.e. anyone using go1.17 with
skaffold debug will still hit this) but it helps us pass the tests on main
branch.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>