Add Dir config field to ko builder
The `Dir` config field supports users who have directory layouts where
the Go sources and `go.mod` file are in subdirectories of the `context`
directory.

The value of `Dir` is a directory path relative to the `context`
directory, and it specifies the directory where `go build` is run.

This change also moves forward some of the additional ko builder flags
in the implementation plan, since the relevant ko PRs have been merged.