feat(watch): Support globstar in dependencies
The Bash 4 style `**` glob syntax for matching an arbitrary number of
directories is already supported in the syntax for manual file sync.

This change adds support for globstar to the `dependencies` sections
(both `paths` and `excludes`) for the Buildpacks builder, the custom
builder, and custom tests. The ko builder will also use this syntax for
file watching.

The doublestar module is already used for file sync, so using it for
dependencies provides consistent behavior for users.

The implementation takes care of preserving existing behavior for other
wildcards (`*`), and for cases where dependencies are specified by
directory path (e.g., `.`).
This behavior is documented in the test cases in `list_test.go`.

**Resolves**: #6029
**Related**: #6041, #6350, #5933

Follow-up work: Update syntax in examples, e.g. `**.go` -> `**/*.go`.